The SDL GPU world renderer: everything the scene draws now goes through the
engine's own delegation chain (cScene::Draw -> Camera::DrawScene ->
object->Draw()) into a renderer class per pipeline, instead of the flat
*RenderSDL bridges the earlier slices used.

  Objects   SDLObject3dxRenderer   3dx models, bump/specular materials, static decor
  Terrain   SDLTileMapRenderer     vMap heightfield, baked per-cell colour, per-material
                                   detail grain, shadow caster + receiver, lightmap
  Water     SDLWaterRenderer       the surface, and its planar reflection
  World     SDLWorldQuadRenderer   sky, coast foam, wave sources, particle sprites,
                                   light columns, circle splines, lightning bolts
  UI        SDLUIRenderer          text, sprites, 2D primitives, camera viewport
  Minimap   SDLMinimapRenderer     map, symbols, view zone -- interleaved into the UI's
                                   run list, since it draws part-way through the UI

Plus the shadow map (objects and terrain both cast and receive, 2x2 PCF) and the
terrain lightmap, and the fixes the work turned up along the way: two heap
corruptions (Color4c widening to 8 bytes on LP64, and A8L8 texture staging
under-allocation), the _splitpath backslash shim, and the DrawStrip landmine
whose empty off-Windows stubs left a wild pointer behind.

+++ Util/XMath/Colors.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "XMath.h"
+#include <cstdint>
 
 class Archive;
 struct Color3c;
@@ -64,7 +65,10 @@ struct Color4c
 {
 	union{
 		struct{ unsigned char b,g,r,a; };
-		struct{ unsigned long argb; };
+		// Must stay 32 bits wide: `unsigned long` is 4 bytes on Windows but 8 on the
+		// LP64 platforms, which would make sizeof(Color4c) 8 and silently widen every
+		// vertex struct holding a diffuse past the stride its D3D declaration states.
+		struct{ uint32_t argb; };
 	};
 	
 	Color4c()										{ }
